% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_foot.R
\name{fit_foot}
\alias{fit_foot}
\title{Fit any football model}
\usage{
fit_foot(data, model)
}
\arguments{
\item{data}{a data frame, or a matrix containing the following mandatory items: home team, away team,
home goals, away goals.}
}
\description{
Fit any football model
}
