% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foot_priors.R
\name{normal}
\alias{normal}
\title{Football priors}
\usage{
normal(location = 0, scale = NULL, autoscale = TRUE)
}
\description{
This prior specification is just a duplicate
of some of the priors used by the \pkg{rstanarm} package. They can be passed to the
\code{stan_foot} function, through the arguments \code{prior} and \code{prior_sd}.
See  the vignette \href{http://mc-stan.org/rstanarm/articles/priors.html}{\emph{Prior
  Distributions for rstanarm Models}} for further details.
You can choose between: code{normal}, \code{cauchy}, \code{laplace}, \code{student_t}.
}
