% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.btdFoot.R
\name{print.btdFoot}
\alias{print.btdFoot}
\title{Print Method for btdFoot Objects}
\usage{
\method{print}{btdFoot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{btdFoot}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Displays a summary of the Bayesian Bradley-Terry-Davidson model fit, including rank measures and a summary of the Stan fit.
}
