% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.stanFoot.R
\name{print.stanFoot}
\alias{print.stanFoot}
\title{Print Method for stanFoot Objects}
\usage{
\method{print}{stanFoot}(x, ...)
}
\arguments{
\item{x}{An object of class \code{stanFoot}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Displays a summary of the Stan football model fit, including posterior summaries of parameters.
}
