% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_foot.R
\name{compare_foot}
\alias{compare_foot}
\title{Compare Football Models using Various Metrics}
\usage{
compare_foot(
  models,
  test_data,
  metric = c("accuracy", "brier", "mcFaddenR2", "coxSnellR2", "ACP")
)
}
\arguments{
\item{models}{A named list of fitted model objects (of class \code{stanFoot} or \code{stanfit}), each representing a football model.}

\item{test_data}{A data frame containing the test dataset, with columns:
\itemize{
  \item \code{home_team}: Home team's name (character string).
  \item \code{away_team}: Away team's name (character string).
  \item \code{homegoals}: Goals scored by the home team (integer >= 0).
  \item \code{awaygoals}: Goals scored by the away team (integer >= 0).
}}

\item{metric}{A character vector specifying the metrics to use for comparison. Options are:
\itemize{
  \item \code{"accuracy"}: Computes the accuracy of each model.
  \item \code{"brier"}: Computes the Brier score of each model.
  \item \code{"mcFaddenR2"}: Computes McFadden's Pseudo $R^2$ for each model.
  \item \code{"coxSnellR2"}: Computes the Cox-Snell Pseudo $R^2$ for each model.
  \item \code{"ACP"}: Computes the Average Coverage Probability (ACP) for each model.
}
Default is \code{c("accuracy", "brier", "mcFaddenR2", "coxSnellR2", "ACP")}, computing the specified metrics.}
}
\value{
A data frame containing the metric values for each model.
}
\description{
Compares multiple football models based on specified metrics (accuracy, Brier score, Cox-Snell Pseudo $R^2$, Average Coverage Probability, and/or McFadden's Pseudo $R^2$), using a test dataset.
}
\details{
The function extracts predictions from each model and computes the chosen metrics on the test dataset.
}
\examples{
\dontrun{

require(dplyr)

data("italy")
italy_2000 <- italy \%>\%
 dplyr::select(Season, home, visitor, hgoal,vgoal) \%>\%
 dplyr::filter(Season=="2000")

fit <- stan_foot(data = italy_2000,
                 model="double_pois", predict =18)  # double pois

italy_2000_test <- italy_2000[289:306,]

colnames(italy_2000_test) <- c("season", "home_team",  "away_team", "homegoals", "awaygoals")

compare_results <- compare_foot(
  models = list(model_1 = fit),
  test_data = italy_2000_test,
  metric = c("accuracy", "brier", "mcFaddenR2", "coxSnellR2", "ACP")
)

print(compare_results)
}
}
