% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foot_plot.R
\name{foot_abilities}
\alias{foot_abilities}
\title{Plots}
\usage{
foot_abilities(object, teams)
}
\arguments{
\item{object}{An object of class \code{stanfit} as given by \code{stan_foot} function.}

\item{teams}{Character vector with the list of teams as extracted from the dataset.}
}
\description{
Plots
}
