% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foot_plot.R
\name{foot_abilities}
\alias{foot_abilities}
\title{Plots}
\usage{
foot_abilities(object, teams)
}
\arguments{
\item{object}{An object of class \code{stanfit} as given by \code{stan_foot} function.}

\item{teams}{Character vector with the list of teams as extracted from the dataset.}
}
\description{
Plots
}
\examples{
library(engsoccerdata)
library(tidyverse)

ristr_italy <- as_data_frame(italy)
ristr_italy<- ristr_italy \%>\%
 select(Season, home, visitor, hgoal,vgoal) \%>\%
 filter(Season=="2000" |  Season=="2001"| Season =="2002")
fit<-stan_foot(data = ristr_italy,
                       model="biv_pois", predict =306,
                       dynamic_type = "seasonal")
teams <- unique(ristr_italy$home)
foot_abilities(fit, teams)

 @export
}
