% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foot_prob.R
\name{foot_prob}
\alias{foot_prob}
\title{Plot football matches probabilities from Stan model}
\usage{
foot_prob(object, data, home_team, away_team)
}
\arguments{
\item{object}{An object of class \code{stanfit} as given by \code{stan_foot} function.}

\item{data}{A data frame, or a matrix containing the following mandatory items: home team, away team,
home goals, away goals.}

\item{home_team}{The home team for the predicted match.}

\item{away_team}{The away team for the predicted match.}
}
\description{
Depicts probabilities from out-of-sample football matches.
}
\examples{
\dontrun{
### weekly dynamics, predict the last four weeks
italy_2000<- italy \%>\%
 dplyr::select(Season, home, visitor, hgoal,vgoal) \%>\%
 filter(Season=="2000")

fit <- stan_foot(data = italy_2000,
                 model="double_pois", predict =18,
                 dynamic_type = "weekly")  # double pois

foot_prob(fit, italy_2000, "Inter",
          "Bologna FC")

foot_prob(fit, italy_2000, "Reggina Calcio",
           "AC Milan")
}
}
