% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foot_prob.R
\name{foot_prob}
\alias{foot_prob}
\title{Probabilities}
\usage{
foot_prob(object, teams, data, home_team, away_team, predict,
  true_gol_home = 0, true_gol_away = 0)
}
\arguments{
\item{object}{}
}
\description{
Probabilities
}
\examples{

### weekly dynamics, predict the last four weeks
italy_2000<- italy \%>\%
 dplyr::select(Season, home, visitor, hgoal,vgoal) \%>\%
 filter(Season=="2000")

fit <- stan_foot(data = italy_2000,
                 model="student_t", predict =18,
                 dynamic_type = "weekly")  # double poisson

teams <- unique(italy_2000$home)
foot_prob(fit, teams, italy_2000, "Inter",
          "Bologna FC", predict = 18)
}
