% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_btdPosterior.R
\name{plot_btdPosterior}
\alias{plot_btdPosterior}
\title{Plot Posterior Distributions for btdFoot Objects}
\usage{
plot_btdPosterior(x, teams_of_interest = NULL, ncol = NULL, scales = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{btdFoot}.}

\item{teams_of_interest}{Optional. A character vector of team names to include in the posterior boxplots. If \code{NULL}, all teams are included.}

\item{ncol}{Optional. An integer specifying the number of columns in the facet wrap. Defaults:
\itemize{
  \item Dynamic Ranking: \code{ncol = 8}.
  \item Static Ranking: \code{ncol = 4}.
}}

\item{scales}{Optional. A character string specifying the scales for the facets. Options:
\itemize{
  \item Dynamic Ranking: \code{"free_x"}, \code{"free_y"}, \code{"free"}.
  \item Static Ranking: \code{"free_y"}, \code{"fixed"}.
}
 Defaults:
\itemize{
  \item Dynamic Ranking: \code{"free_x"}.
  \item Static Ranking: \code{"free_x"}.
}}

\item{...}{Additional arguments passed to \code{geom_boxplot()} for customization (e.g., \code{size}, \code{alpha}, \code{color}).}
}
\value{
A ggplot object representing the posterior distributions plot.
}
\description{
Visualizes posterior distributions of team strengths with customizable facets.
}
\details{
\itemize{
\item Dynamic Ranking: Faceted boxplots of posterior log-strengths by team and period.
\item Static Ranking: Faceted boxplots of posterior log-strengths for each team.

}
}
