% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_logStrength.R
\name{plot_logStrength}
\alias{plot_logStrength}
\title{Plot Rankings for btdFoot Objects}
\usage{
plot_logStrength(x, teams_of_interest = NULL, ncol = 8, ...)
}
\arguments{
\item{x}{An object of class \code{btdFoot}.}

\item{teams_of_interest}{Optional. A character vector of team names to include in the rankings plot. If \code{NULL}, all teams are included.}

\item{ncol}{Optional. An integer specifying the number of columns in the facet wrap for dynamic ranking plots. Defaults to \code{8}.}

\item{...}{Additional arguments passed to \code{geom_line()}, \code{geom_point()}, and \code{geom_segment()} for customization (e.g., \code{size}, \code{alpha}, \code{color}).}
}
\value{
A ggplot object representing the rankings plot.
}
\description{
Visualizes team rankings based on whether the ranking is dynamic or static.
}
\details{
- **Dynamic Ranking**: Plots Rank Points over Periods for each team with lines and points.
- **Static Ranking**: Plots Rank Points on the x-axis against Team Names on the y-axis with horizontal lines and points.
}
\author{
Roberto Macrì Demartino \email{roberto.macridemartino@phd.unipd.it}.
}
