% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_foot.R
\name{pp_foot}
\alias{pp_foot}
\title{Posterior predictive checks for football  models}
\usage{
pp_foot(data, object, type = c("aggregated", "matches"))
}
\description{
Posterior predictive checks for football  models
}
