% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_foot.R
\name{pp_foot}
\alias{pp_foot}
\title{Posterior predictive checks for football models}
\usage{
pp_foot(data, object, type = c("aggregated", "matches"), coverage = 0.95)
}
\arguments{
\item{data}{A data frame, or a matrix containing the following mandatory items: home team, away team,
home goals, away goals.}

\item{object}{An object of class \code{\link[rstan]{stanfit}} as given by \code{stan_foot} function.}

\item{type}{Type of plots, one among \code{"aggregated"} or \code{"matches"}.}

\item{coverage}{Argument to specify the width \eqn{1-\alpha} of posterior probability intervals. Default is 0.95.}
}
\value{
Posterior predictive plots: when \code{"aggregated"} (default) is selected, the function
returns a frequency plot for some pre-selected goal-difference values,
along with their correspondent Bayesian p-values, computed as
\eqn{Pr(y_rep \ge y)|y)}, where \eqn{y_rep} is a data replication from the
posterior predictive distribution (more details in Gelman et al., 2013).
Bayesian p-values very close to 0 or 1 could exhibit
possible model misfits.

When \code{"matches"} is selected an ordered-frequency plot for all the
goal-differences in the considered matches is provided, along with the
empirical Bayesian coverage at level \eqn{1-\alpha}.
}
\description{
The function provides posterior predictive plots to check the adequacy of the Bayesian models as
returned by the \code{stan_foot} function.
}
\examples{

\donttest{
require(dplyr)

data("italy")
italy_2000<- italy \%>\%
 dplyr::select(Season, home, visitor, hgoal,vgoal) \%>\%
 dplyr::filter(Season=="2000")

fit <- stan_foot(italy_2000, "double_pois", iter = 200)

pp_foot(italy_2000, fit)

}

}
\references{
Gelman, A., Carlin, J. B., Stern, H. S., Dunson, D. B., Vehtari, A., & Rubin, D. B. (2013). Bayesian data analysis. CRC press.
}
\author{
Leonardo Egidi \email{legidi@units.it}
}
