% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fun_foot.R
\name{print.stanFoot}
\alias{print.stanFoot}
\title{Print Method for stanFoot Objects}
\usage{
\method{print}{stanFoot}(x, pars = NULL, teams = NULL, digits = 3, true_names = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{stanFoot}.}

\item{pars}{Optional character vector specifying parameters to include in the summary. This can be specific parameter names or macro parameter groups (e.g., \code{"att"}, \code{"def"}, \code{"att_raw"}, \code{"def_raw"}, \code{"home"}, \code{"sigma_att"}, \code{"sigma_def"}, \code{"Sigma_att"}, \code{"Sigma_def"}, \code{"rho"}, \code{"beta"}, and \code{"sigma_y"}). If \code{NULL}, all parameters are included.}

\item{teams}{Optional character vector specifying team names whose \code{"att"}, \code{"def"}, \code{"att_raw"}, \code{"def_raw"} parameters should be displayed.}

\item{digits}{Number of significant digits to use when printing numeric values.}

\item{true_names}{Logical value indicating whether to display team names in parameter summaries. Default is \code{TRUE}.}

\item{...}{Additional arguments passed.}
}
\description{
Provides detailed posterior summaries for the Stan football model parameters.
}
