% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_foot.R
\name{stan_foot}
\alias{stan_foot}
\title{Fit football models  with Stan}
\usage{
stan_foot(data, model, predict, n.iter = 200, chains = 4,
  dynamic_type = FALSE)
}
\arguments{
\item{data}{A data frame, or a matrix containing the following mandatory items: home team, away team,
home goals, away goals.}

\item{model}{The type of Stan model used to fit the data.
One among the following: \code{"double_pois"},
\code{"biv_pois"}, \code{"skellam"}, \code{"student_t"}.}

\item{predict}{The number of out-of-sample matches.}
}
\description{
Stan football modelling for the most famous models:
double Poisson, bivariate Poisson, Skellam and student t.
}
\examples{
library(engsoccerdata)
library(tidyverse)

### Use Italian Serie A from 2000 to 2002

italy <- as_tibble(italy)
italy_2000_2002<- italy \%>\%
 dplyr::select(Season, home, visitor, hgoal,vgoal) \%>\%
 filter(Season=="2000" |  Season=="2001"| Season=="2002")

### Fit Stan models
## no dynamics, no predictions

fit1 <- stan_foot(data = italy_2000_2002,
                model="double_pois") # double poisson
print(fit1, pars =c("home", "sigma_att",
                    "sigma_def"))

fit2 <- stan_foot(data = italy_2000_2002,
                model="biv_pois")    # bivariate poisson
print(fit2, pars =c("home", "rho"
                    "sigma_att", "sigma_def"))

fit3 <- stan_foot(data = italy_2000_2002,
                model="skellam")     # skellam
print(fit3, pars =c("home", "sigma_att",
                    "sigma_def"))

fit4 <- stan_foot(data = italy_2000_2002,
                model="student_t")   # student_t
print(fit4, pars =c("home", "beta")

## seasonal dynamics, no prediction

fit5 <- stan_foot(data = italy_2000_2002,
                model="double_pois",
                dynamic_type ="seasonal") # double poisson
print(fit5, pars =c("home", "Sigma_att",
                    "Sigma_def"))

## seasonal dynamics, prediction for the last season

fit6 <- stan_foot(data = italy_2000_2002,
                model="double_pois",
                dynamic_type ="seasonal",
                predict = 306) # double poisson
print(fit6, pars =c("home", "Sigma_att",
                    "Sigma_def"))
}
