% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_foot.R
\name{stan_foot}
\alias{stan_foot}
\title{Fit any football model}
\usage{
stan_foot(data, model, predict, n.iter = 200, chains = 4,
  dynamic_type = FALSE)
}
\arguments{
\item{data}{a data frame, or a matrix containing the following mandatory items: home team, away team,
home goals, away goals.}

\item{predict}{}
}
\description{
Fit any football model
}
\examples{
library(engsoccerdata)
library(tidyverse)

ristr_italy <- as_data_frame(italy)
ristr_italy<- ristr_italy \%>\%
 select(Season, home, visitor, hgoal,vgoal) \%>\%
 filter(Season=="2000" |  Season=="2001")
}
