% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.btdFoot.R
\name{summary.btdFoot}
\alias{summary.btdFoot}
\title{Summary Method for btdFoot Objects}
\usage{
\method{summary}{btdFoot}(object, ...)
}
\arguments{
\item{object}{An object of class \code{btdFoot}.}

\item{...}{Additional arguments.}
}
\description{
Provides detailed posterior summaries for the Bayesian Bradley-Terry-Davidson model parameters.
}
