% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.stanFoot.R
\name{summary.stanFoot}
\alias{summary.stanFoot}
\title{Summary Method for stanFoot Objects}
\usage{
\method{summary}{stanFoot}(object, ...)
}
\arguments{
\item{object}{An object of class \code{stanFoot}.}

\item{...}{Additional arguments.}
}
\description{
Provides detailed posterior summaries for the Stan football model parameters.
}
